move widget to core
