+++ ephios/core/widgets.py
@@ -1,3 +1,4 @@
+from django.forms import Widget
 from django_select2.forms import ModelSelect2MultipleWidget, ModelSelect2Widget
 
 from ephios.core.models import UserProfile
@@ -17,3 +18,16 @@ class MultiUserProfileWidget(ModelSelect2MultipleWidget):
 
     def label_from_instance(self, obj):
         return obj.get_full_name()
+
+
+class PreviousCommentWidget(Widget):
+    template_name = "core/widgets/previous_comments.html"
+
+    def __init__(self, *args, **kwargs):
+        self.comments = kwargs.pop("comments")
+        super().__init__(*args, **kwargs)
+
+    def get_context(self, name, value, attrs):
+        context = super().get_context(name, value, attrs)
+        context["comments"] = self.comments
+        return context
